refactor(blockifier): move ugly code to native blockifier
